WIP - KumoParameters with multi asset
@wollooo and @asimpk, please regularly sync your work with this feature branch

+++ packages/contracts/contracts/PriceFeed.sol
@@ -38,24 +38,24 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
 
     bytes32 constant public ETHUSD_TELLOR_REQ_ID = 0x0000000000000000000000000000000000000000000000000000000000000001;
 
-    // Use to convert a price answer to an 18-digit precision uint
-    uint constant public TARGET_DIGITS = 18;  
-    uint constant public TELLOR_DIGITS = 6;
+    // Use to convert a price answer to an 18-digit precision uint256
+    uint256 constant public TARGET_DIGITS = 18;  
+    uint256 constant public TELLOR_DIGITS = 6;
 
     // Maximum time period allowed since Chainlink's latest round data timestamp, beyond which Chainlink is considered frozen.
-    uint constant public TIMEOUT = 14400;  // 4 hours: 60 * 60 * 4
+    uint256 constant public TIMEOUT = 14400;  // 4 hours: 60 * 60 * 4
     
     // Maximum deviation allowed between two consecutive Chainlink oracle prices. 18-digit precision.
-    uint constant public MAX_PRICE_DEVIATION_FROM_PREVIOUS_ROUND =  5e17; // 50%
+    uint256 constant public MAX_PRICE_DEVIATION_FROM_PREVIOUS_ROUND =  5e17; // 50%
 
     /* 
     * The maximum relative price difference between two oracle responses allowed in order for the PriceFeed
     * to return to using the Chainlink oracle. 18-digit precision.
     */
-    uint constant public MAX_PRICE_DIFFERENCE_BETWEEN_ORACLES = 5e16; // 5%
+    uint256 constant public MAX_PRICE_DIFFERENCE_BETWEEN_ORACLES = 5e16; // 5%
 
     // The last good price seen from an oracle by Kumo
-    uint public lastGoodPrice;
+    uint256 public lastGoodPrice;
 
     struct ChainlinkResponse {
         uint80 roundId;
@@ -83,7 +83,7 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
     // The current status of the PricFeed, which determines the conditions for the next price fetch attempt
     Status public status;
 
-    // event LastGoodPriceUpdated(uint _lastGoodPrice);
+    // event LastGoodPriceUpdated(uint256 _lastGoodPrice);
     event PriceFeedStatusChanged(Status newStatus);
 
     // --- Dependency setters ---
@@ -133,7 +133,7 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
     * it uses the last good price seen by Kumo.
     *
     */
-    function fetchPrice() external override returns (uint latestprice_) {
+    function fetchPrice() external override returns (uint256 latestprice_) {
         // Get current and previous price data from Chainlink, and current price data from Tellor
         ChainlinkResponse memory chainlinkResponse = _getCurrentChainlinkResponse();
         ChainlinkResponse memory prevChainlinkResponse = _getPrevChainlinkResponse(chainlinkResponse.roundId, chainlinkResponse.decimals);
@@ -373,18 +373,18 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
     }
 
     function _chainlinkPriceChangeAboveMax(ChainlinkResponse memory _currentResponse, ChainlinkResponse memory _prevResponse) internal pure returns (bool) {
-        uint currentScaledPrice = _scaleChainlinkPriceByDigits(uint256(_currentResponse.answer), _currentResponse.decimals);
-        uint prevScaledPrice = _scaleChainlinkPriceByDigits(uint256(_prevResponse.answer), _prevResponse.decimals);
+        uint256 currentScaledPrice = _scaleChainlinkPriceByDigits(uint256(_currentResponse.answer), _currentResponse.decimals);
+        uint256 prevScaledPrice = _scaleChainlinkPriceByDigits(uint256(_prevResponse.answer), _prevResponse.decimals);
 
-        uint minPrice = KumoMath._min(currentScaledPrice, prevScaledPrice);
-        uint maxPrice = KumoMath._max(currentScaledPrice, prevScaledPrice);
+        uint256 minPrice = KumoMath._min(currentScaledPrice, prevScaledPrice);
+        uint256 maxPrice = KumoMath._max(currentScaledPrice, prevScaledPrice);
 
         /*
         * Use the larger price as the denominator:
         * - If price decreased, the percentage deviation is in relation to the the previous price.
         * - If price increased, the percentage deviation is in relation to the current price.
         */
-        uint percentDeviation = maxPrice.sub(minPrice).mul(DECIMAL_PRECISION).div(maxPrice);
+        uint256 percentDeviation = maxPrice.sub(minPrice).mul(DECIMAL_PRECISION).div(maxPrice);
 
         // Return true if price has more than doubled, or more than halved.
         return percentDeviation > MAX_PRICE_DEVIATION_FROM_PREVIOUS_ROUND;
@@ -431,13 +431,13 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
     }
 
     function _bothOraclesSimilarPrice( ChainlinkResponse memory _chainlinkResponse, TellorResponse memory _tellorResponse) internal pure returns (bool) {
-        uint scaledChainlinkPrice = _scaleChainlinkPriceByDigits(uint256(_chainlinkResponse.answer), _chainlinkResponse.decimals);
-        uint scaledTellorPrice = _scaleTellorPriceByDigits(_tellorResponse.value);
+        uint256 scaledChainlinkPrice = _scaleChainlinkPriceByDigits(uint256(_chainlinkResponse.answer), _chainlinkResponse.decimals);
+        uint256 scaledTellorPrice = _scaleTellorPriceByDigits(_tellorResponse.value);
 
         // Get the relative price difference between the oracles. Use the lower price as the denominator, i.e. the reference for the calculation.
-        uint minPrice = KumoMath._min(scaledTellorPrice, scaledChainlinkPrice);
-        uint maxPrice = KumoMath._max(scaledTellorPrice, scaledChainlinkPrice);
-        uint percentPriceDifference = maxPrice.sub(minPrice).mul(DECIMAL_PRECISION).div(minPrice);
+        uint256 minPrice = KumoMath._min(scaledTellorPrice, scaledChainlinkPrice);
+        uint256 maxPrice = KumoMath._max(scaledTellorPrice, scaledChainlinkPrice);
+        uint256 percentPriceDifference = maxPrice.sub(minPrice).mul(DECIMAL_PRECISION).div(minPrice);
 
         /*
         * Return true if the relative price difference is <= 3%: if so, we assume both oracles are probably reporting
@@ -446,14 +446,14 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
         return percentPriceDifference <= MAX_PRICE_DIFFERENCE_BETWEEN_ORACLES;
     }
 
-    function _scaleChainlinkPriceByDigits(uint _price, uint _answerDigits) internal pure returns (uint) {
+    function _scaleChainlinkPriceByDigits(uint256 _price, uint256 _answerDigits) internal pure returns (uint256) {
         /*
         * Convert the price returned by the Chainlink oracle to an 18-digit decimal for use by Kumo.
         * At date of Kumo launch, Chainlink uses an 8-digit price, but we also handle the possibility of
         * future changes.
         *
         */
-        uint price;
+        uint256 price;
         if (_answerDigits >= TARGET_DIGITS) {
             // Scale the returned price value down to Kumo's target precision
             price = _price.div(10 ** (_answerDigits - TARGET_DIGITS));
@@ -465,7 +465,7 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
         return price;
     }
 
-    function _scaleTellorPriceByDigits(uint _price) internal pure returns (uint) {
+    function _scaleTellorPriceByDigits(uint256 _price) internal pure returns (uint256) {
         return _price.mul(10**(TARGET_DIGITS - TELLOR_DIGITS));
     }
 
@@ -474,20 +474,20 @@ contract PriceFeed is Ownable, CheckContract, BaseMath, IPriceFeed {
         emit PriceFeedStatusChanged(_status);
     }
 
-    function _storePrice(uint _currentPrice) internal {
+    function _storePrice(uint256 _currentPrice) internal {
         lastGoodPrice = _currentPrice;
         emit LastGoodPriceUpdated(_currentPrice);
     }
 
-     function _storeTellorPrice(TellorResponse memory _tellorResponse) internal returns (uint) {
-        uint scaledTellorPrice = _scaleTellorPriceByDigits(_tellorResponse.value);
+     function _storeTellorPrice(TellorResponse memory _tellorResponse) internal returns (uint256) {
+        uint256 scaledTellorPrice = _scaleTellorPriceByDigits(_tellorResponse.value);
         _storePrice(scaledTellorPrice);
 
         return scaledTellorPrice;
     }
 
-    function _storeChainlinkPrice(ChainlinkResponse memory _chainlinkResponse) internal returns (uint) {
-        uint scaledChainlinkPrice = _scaleChainlinkPriceByDigits(uint256(_chainlinkResponse.answer), _chainlinkResponse.decimals);
+    function _storeChainlinkPrice(ChainlinkResponse memory _chainlinkResponse) internal returns (uint256) {
+        uint256 scaledChainlinkPrice = _scaleChainlinkPriceByDigits(uint256(_chainlinkResponse.answer), _chainlinkResponse.decimals);
         _storePrice(scaledChainlinkPrice);
 
         return scaledChainlinkPrice;
